Ensure that installed marker is installed before enrollment.
