--- internal/pkg/agent/application/info/state.go
@@ -1,21 +1,0 @@
-// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
-// or more contributor license agreements. Licensed under the Elastic License;
-// you may not use this file except in compliance with the Elastic License.
-
-package info
-
-import (
-	"os"
-	"path/filepath"
-
-	"github.com/elastic/elastic-agent/internal/pkg/agent/application/paths"
-	"github.com/elastic/elastic-agent/pkg/utils"
-)
-
-func CreateInstallMarker(topPath string, ownership utils.FileOwner) error {
-	markerFilePath := filepath.Join(topPath, paths.MarkerFileName)
-	if _, err := os.Create(markerFilePath); err != nil {
-		return err
-	}
-	return fixInstallMarkerPermissions(markerFilePath, ownership)
-}
